Do not spawn a new process for sending data to coveralls
Use the handleInput function of the coveralls module directly to
avoid problems with waiting on child processes. It seams that
Grunt ignores the call to async() and does not wait for finished
tasks before exit. If submitting data to coveralls takes longer
and the task is the last one, it is possible that the process
exits before all data has been submitted.